feat(Drawing): add functionality to create a new drawing with selected modules and fields
This commit adds functionality to create a new drawing with selected modules
and fields. The user can select the modules and fields to include in the drawing
and set a title for it. The selected modules and fields are stored in the state
and used to create the drawing when the user clicks the "Agregar" button. The
drawing is then added to the list of drawings in the context.